+++ .mise.toml
@@ -1,0 +1,5 @@
+[tools]
+python = "3.10"
+
+[env]
+_.python.venv = ".venv"
